requirements.txt
fastapi==0.118.0
uvicorn[standard]==0.30.0
asyncpg==0.30.0
aioredis==2.0.1
redis==6.4.0
python-dotenv==1.1.1
PyJWT==2.10.1
pydantic==2.11.9
